proofs only imports from certifiers, not lc
