+++ certifiers/errors_test.go
@@ -1,4 +1,4 @@
-package lightclient
+package certifiers
 
 import (
 	"errors"
@@ -15,14 +15,4 @@ func TestErrorHeight(t *testing.T) {
 	e2 := errors.New("foobar")
 	assert.False(t, IsHeightMismatchErr(e2))
 	assert.False(t, IsHeightMismatchErr(nil))
-}
-
-func TestErrorNoData(t *testing.T) {
-	e1 := ErrNoData()
-	e1.Error()
-	assert.True(t, IsNoDataErr(e1))
-
-	e2 := errors.New("foobar")
-	assert.False(t, IsNoDataErr(e2))
-	assert.False(t, IsNoDataErr(nil))
 }
