+++ proofs/app_test.go
@@ -6,23 +6,26 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	lc "github.com/tendermint/light-client"
-	"github.com/tendermint/light-client/proofs"
-	"github.com/tendermint/tendermint/rpc/client"
+
 	ctest "github.com/tendermint/tmlibs/test"
+
+	"github.com/tendermint/tendermint/rpc/client"
+
+	"github.com/tendermint/light-client/certifiers"
+	"github.com/tendermint/light-client/proofs"
 )
 
-func getCurrentCheck(t *testing.T, cl client.Client) lc.Checkpoint {
+func getCurrentCheck(t *testing.T, cl client.Client) certifiers.Checkpoint {
 	stat, err := cl.Status()
 	require.Nil(t, err, "%+v", err)
 	return getCheckForHeight(t, cl, stat.LatestBlockHeight)
 }
 
-func getCheckForHeight(t *testing.T, cl client.Client, h int) lc.Checkpoint {
+func getCheckForHeight(t *testing.T, cl client.Client, h int) certifiers.Checkpoint {
 	client.WaitForHeight(cl, h, nil)
 	commit, err := cl.Commit(&h)
 	require.Nil(t, err, "%+v", err)
-	return lc.CheckpointFromResult(commit)
+	return certifiers.CheckpointFromResult(commit)
 }
 
 func TestAppProofs(t *testing.T) {
@@ -74,7 +77,7 @@ func TestAppProofs(t *testing.T) {
 // and validate with the checkpoint.  It also does lots of modifications
 // to the binary data and makes sure no mods validates properly
 func testSerialization(t *testing.T, prover proofs.Prover, pr proofs.Proof,
-	check lc.Checkpoint, mods int) {
+	check certifiers.Checkpoint, mods int) {
 
 	require := require.New(t)
