+++ proofs/block.go
@@ -4,16 +4,16 @@ import (
 	"bytes"
 	"errors"
 
-	lc "github.com/tendermint/light-client"
+	"github.com/tendermint/light-client/certifiers"
 	"github.com/tendermint/tendermint/types"
 )
 
-func ValidateBlockMeta(meta *types.BlockMeta, check lc.Checkpoint) error {
+func ValidateBlockMeta(meta *types.BlockMeta, check certifiers.Checkpoint) error {
 	// TODO: check the BlockID??
 	return ValidateHeader(meta.Header, check)
 }
 
-func ValidateBlock(meta *types.Block, check lc.Checkpoint) error {
+func ValidateBlock(meta *types.Block, check certifiers.Checkpoint) error {
 	err := ValidateHeader(meta.Header, check)
 	if err != nil {
 		return err
@@ -24,10 +24,10 @@ func ValidateBlock(meta *types.Block, check lc.Checkpoint) error {
 	return nil
 }
 
-func ValidateHeader(head *types.Header, check lc.Checkpoint) error {
+func ValidateHeader(head *types.Header, check certifiers.Checkpoint) error {
 	// make sure they are for the same height (obvious fail)
 	if head.Height != check.Height() {
-		return lc.ErrHeightMismatch(head.Height, check.Height())
+		return certifiers.ErrHeightMismatch(head.Height, check.Height())
 	}
 	// check if they are equal by using hashes
 	chead := check.Header
