+++ proofs/wrapper.go
@@ -4,12 +4,13 @@ import (
 	"fmt"
 
 	"github.com/tendermint/go-wire/data"
-	lc "github.com/tendermint/light-client"
-	"github.com/tendermint/light-client/certifiers"
+	"github.com/tendermint/tmlibs/events"
+
 	rpcclient "github.com/tendermint/tendermint/rpc/client"
 	ctypes "github.com/tendermint/tendermint/rpc/core/types"
 	"github.com/tendermint/tendermint/types"
-	"github.com/tendermint/tmlibs/events"
+
+	"github.com/tendermint/light-client/certifiers"
 )
 
 var _ rpcclient.Client = Wrapper{}
@@ -56,7 +57,7 @@ func (w Wrapper) proveQuery(r *ctypes.ResultABCIQuery) (*ctypes.ResultABCIQuery,
 		return nil, err
 	}
 	// make sure the checkpoint and proof match up
-	check := lc.CheckpointFromResult(c)
+	check := certifiers.CheckpointFromResult(c)
 	// verify query
 	proof := AppProof{
 		Height: r.Height,
@@ -79,7 +80,7 @@ func (w Wrapper) Tx(hash []byte, prove bool) (*ctypes.ResultTx, error) {
 		return nil, err
 	}
 	// make sure the checkpoint and proof match up
-	check := lc.CheckpointFromResult(c)
+	check := certifiers.CheckpointFromResult(c)
 	// verify tx
 	proof := TxProof{
 		Height: uint64(r.Height),
@@ -102,7 +103,7 @@ func (w Wrapper) BlockchainInfo(minHeight, maxHeight int) (*ctypes.ResultBlockch
 		if err != nil {
 			return nil, err
 		}
-		check := lc.CheckpointFromResult(c)
+		check := certifiers.CheckpointFromResult(c)
 		err = ValidateBlockMeta(meta, check)
 		if err != nil {
 			return nil, err
@@ -122,7 +123,7 @@ func (w Wrapper) Block(height *int) (*ctypes.ResultBlock, error) {
 	if err != nil {
 		return nil, err
 	}
-	check := lc.CheckpointFromResult(c)
+	check := certifiers.CheckpointFromResult(c)
 
 	// now verify
 	err = ValidateBlockMeta(r.BlockMeta, check)
@@ -144,7 +145,7 @@ func (w Wrapper) Commit(height *int) (*ctypes.ResultCommit, error) {
 	r, err := w.Client.Commit(height)
 	// if we got it, then certify it
 	if err == nil {
-		check := lc.CheckpointFromResult(r)
+		check := certifiers.CheckpointFromResult(r)
 		err = w.cert.Certify(check)
 	}
 	return r, err
@@ -188,7 +189,7 @@ func verifyHeader(c rpcclient.Client, head *types.Header) error {
 	if err != nil {
 		return err
 	}
-	check := lc.CheckpointFromResult(commit)
+	check := certifiers.CheckpointFromResult(commit)
 	return ValidateHeader(head, check)
 }
 
@@ -198,6 +199,6 @@ func verifyBlock(c rpcclient.Client, block *types.Block) error {
 	if err != nil {
 		return err
 	}
-	check := lc.CheckpointFromResult(commit)
+	check := certifiers.CheckpointFromResult(commit)
 	return ValidateBlock(block, check)
 }
